Handle optional Japanese yomigana values
We encountered Google contacts which had junk data in the optional
yomi fields, which then caused a crash in our program because rather
than a string, the name methdods would return a hash-like object. If
someone needs to access the yomigana fields for real Yomigana Japense
reading aid data, they could always access it via the hash syntax for
the GoogleContactsApi::Contact instance.